+++ pairecservice-20221213/alibabacloud_pairecservice20221213/models/_update_abmetric_request.py
@@ -7,11 +7,16 @@ from darabonba.model import DaraModel
 class UpdateABMetricRequest(DaraModel):
     def __init__(
         self,
+        aggregation_by_user: bool = None,
         definition: str = None,
+        denominator: str = None,
         description: str = None,
         instance_id: str = None,
+        is_binomial_distribution: bool = None,
         left_metric_id: str = None,
         name: str = None,
+        need_significance: bool = None,
+        numerator: str = None,
         operator: str = None,
         realtime: bool = None,
         result_resource_id: str = None,
@@ -21,15 +26,20 @@ class UpdateABMetricRequest(DaraModel):
         table_meta_id: str = None,
         type: str = None,
     ):
+        self.aggregation_by_user = aggregation_by_user
         # This parameter is required.
         self.definition = definition
+        self.denominator = denominator
         # This parameter is required.
         self.description = description
         # This parameter is required.
         self.instance_id = instance_id
+        self.is_binomial_distribution = is_binomial_distribution
         self.left_metric_id = left_metric_id
         # This parameter is required.
         self.name = name
+        self.need_significance = need_significance
+        self.numerator = numerator
         self.operator = operator
         # This parameter is required.
         self.realtime = realtime
@@ -51,8 +61,14 @@ class UpdateABMetricRequest(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.aggregation_by_user is not None:
+            result['AggregationByUser'] = self.aggregation_by_user
+
         if self.definition is not None:
             result['Definition'] = self.definition
+
+        if self.denominator is not None:
+            result['Denominator'] = self.denominator
 
         if self.description is not None:
             result['Description'] = self.description
@@ -60,11 +76,20 @@ class UpdateABMetricRequest(DaraModel):
         if self.instance_id is not None:
             result['InstanceId'] = self.instance_id
 
+        if self.is_binomial_distribution is not None:
+            result['IsBinomialDistribution'] = self.is_binomial_distribution
+
         if self.left_metric_id is not None:
             result['LeftMetricId'] = self.left_metric_id
 
         if self.name is not None:
             result['Name'] = self.name
+
+        if self.need_significance is not None:
+            result['NeedSignificance'] = self.need_significance
+
+        if self.numerator is not None:
+            result['Numerator'] = self.numerator
 
         if self.operator is not None:
             result['Operator'] = self.operator
@@ -94,8 +119,14 @@ class UpdateABMetricRequest(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('AggregationByUser') is not None:
+            self.aggregation_by_user = m.get('AggregationByUser')
+
         if m.get('Definition') is not None:
             self.definition = m.get('Definition')
+
+        if m.get('Denominator') is not None:
+            self.denominator = m.get('Denominator')
 
         if m.get('Description') is not None:
             self.description = m.get('Description')
@@ -103,11 +134,20 @@ class UpdateABMetricRequest(DaraModel):
         if m.get('InstanceId') is not None:
             self.instance_id = m.get('InstanceId')
 
+        if m.get('IsBinomialDistribution') is not None:
+            self.is_binomial_distribution = m.get('IsBinomialDistribution')
+
         if m.get('LeftMetricId') is not None:
             self.left_metric_id = m.get('LeftMetricId')
 
         if m.get('Name') is not None:
             self.name = m.get('Name')
+
+        if m.get('NeedSignificance') is not None:
+            self.need_significance = m.get('NeedSignificance')
+
+        if m.get('Numerator') is not None:
+            self.numerator = m.get('Numerator')
 
         if m.get('Operator') is not None:
             self.operator = m.get('Operator')